chore: refactor new folder suggestions code
